Convert routable into a gem so it becomes a dependency of the whole application
This is an interesting dependency case. Any package in the app that needs the main app to boot depends on Routable whether or not it depends on the package / gem directly to do routing. That is, because it depends on the main application to be bootable, which it isn’t when this package is deleted. One could take this multiple ways:

* Every package should depend on the root package (which depends on Routable) because they all need a runtime to do anything… but then, ideally, the root package wouldn’t depend on any packages to not create cycles. Which we currently do via two ways… 

* Root depends on all packages (to do Routable stuff). And, root has some initializers left, which also create dependencies. The initializers could be moved out. The Routable stuff could be moved into a gem (what this PR does) and thus take it out of the app-package-loading considerations

* packages should implement an interface...